Improve reusability of field select component.
