+++ graylog2-web-interface/src/views/components/widgets/FieldsConfiguration.tsx
@@ -50,6 +50,7 @@ type Props = {
   showDeSelectAll?: boolean,
   showListCollapseButton?: boolean
   showUnit?: boolean,
+  fieldSelect?: React.ComponentType<React.ComponentProps<typeof FieldSelect>>
 }
 
 const FieldsConfiguration = ({
@@ -65,6 +66,7 @@ const FieldsConfiguration = ({
   showDeSelectAll = false,
   showListCollapseButton = false,
   showUnit = false,
+  fieldSelect: FieldSelectComponent = FieldSelect,
 }: Props) => {
   const [showSelectedList, setShowSelectedList] = useState(true);
   const onAddField = useCallback((newField: string) => (
@@ -106,22 +108,22 @@ const FieldsConfiguration = ({
                           onChange={onChange}
                           showUnit={showUnit} />
       )}
-      <FieldSelect id="field-create-select"
-                   onChange={onAddField}
-                   clearable={false}
-                   isFieldQualified={isFieldQualified}
-                   persistSelection={false}
-                   name="field-create-select"
-                   value={undefined}
-                   size={selectSize}
-                   menuPortalTarget={menuPortalTarget}
-                   excludedFields={selectedFields ?? []}
-                   placeholder={createSelectPlaceholder}
-                   ariaLabel={createSelectPlaceholder}
-                   onSelectAllRest={showSelectAllRest && onSelectAllRest}
-                   showSelectAllRest={showSelectAllRest}
-                   onDeSelectAll={onDeselectAll}
-                   showDeSelectAll={showDeSelectAll && !!selectedFields.length} />
+      <FieldSelectComponent id="field-create-select"
+                            onChange={onAddField}
+                            clearable={false}
+                            isFieldQualified={isFieldQualified}
+                            persistSelection={false}
+                            name="field-create-select"
+                            value={undefined}
+                            size={selectSize}
+                            menuPortalTarget={menuPortalTarget}
+                            excludedFields={selectedFields ?? []}
+                            placeholder={createSelectPlaceholder}
+                            ariaLabel={createSelectPlaceholder}
+                            onSelectAllRest={showSelectAllRest && onSelectAllRest}
+                            showSelectAllRest={showSelectAllRest}
+                            onDeSelectAll={onDeselectAll}
+                            showDeSelectAll={showDeSelectAll && !!selectedFields.length} />
     </>
   );
 };
